backend/requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
regex==2020.11.11
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.1
Werkzeug==1.0.1
